+++ src/EfCore/EfCore.Extensions.Tests/Fixtures/MemoryFixture.cs
@@ -1,24 +1,19 @@
 namespace EfCore.Extensions.Tests.Fixtures;
 
-public class SqlServerFixture : IAsyncLifetime
+public class MemoryFixture : IAsyncLifetime
 {
-    private readonly MsSqlContainer _sql = new MsSqlBuilder()
-        .WithImage("mcr.microsoft.com/mssql/server:2022-latest")
-        .Build();
-
     public MyDbContext? Db { get; private set; }
 
     public async Task InitializeAsync()
     {
-        await _sql.StartAsync();
-        await Task.Delay(TimeSpan.FromSeconds(5));
+        await Task.Delay(TimeSpan.FromSeconds(5), CancellationToken.None);
 
         var options = new DbContextOptionsBuilder()
             .LogTo(Console.WriteLine, (eventId, logLevel) => logLevel >= LogLevel.Information
                                                              || eventId == RelationalEventId.CommandExecuting)
             .EnableSensitiveDataLogging()
             .EnableDetailedErrors()
-            .UseSqlServer(_sql.GetConnectionString())
+            .UseInMemoryDatabase("testDb")
             .UseAutoConfigModel(op => op.ScanFrom(typeof(MyDbContext).Assembly))
             //DONOT use auto seeding here as there are a dedicated test for it
             //.UseAutoDataSeeding()
@@ -32,11 +27,5 @@ public class SqlServerFixture : IAsyncLifetime
     {
         if (Db != null)
             await Db.DisposeAsync();
-        await _sql.StopAsync();
-        await _sql.DisposeAsync();
     }
-
-    public string GetConnectionString(string dbName) =>
-        _sql.GetConnectionString()
-            .Replace("Database=master;", $"Database={dbName};", StringComparison.OrdinalIgnoreCase);
 }
